+++ banner/src/main/AndroidManifest.xml
