+++ banner/src/main/java/com/ms/banner/BannerScroller.java
