+++ banner/src/main/java/com/ms/banner/holder/BannerViewHolder.java
@@ -5,13 +5,7 @@ import android.view.View;
 
 public interface BannerViewHolder<T> {
 
-    /**
-     * 创建View
-     */
     View createView(Context context);
 
-    /**
-     * 绑定数据
-     */
     void onBind(Context context, int position, T data);
 }
