+++ banner/src/main/java/com/ms/banner/listener/OnBannerClickListener.java
@@ -2,8 +2,5 @@ package com.ms.banner.listener;
 
 public interface OnBannerClickListener {
 
-    /**
-     * 点击事件
-     */
     void onBannerClick(int position);
 }
