+++ banner/src/main/java/com/ms/banner/transformer/ScaleRightTransformer.java
