+++ banner/src/main/java/com/ms/banner/view/ArcShapeView.java
