+++ banner/src/main/res/layout/banner.xml
