+++ banner/src/main/res/values/attr.xml
