+++ banner/src/main/res/values/ids.xml
