+++ settings.gradle
@@ -1,2 +1,1 @@
-include ':rollbanner',':app'
-rootProject.name = 'Banner'
+include ':banner',':app'
